Ensure only rank-0 handles writing files in ddp training

+++ sleap_nn/training/model_trainer.py
@@ -5,6 +5,7 @@ import os
 import shutil
 import copy
 import attrs
+import torch
 import sleap_io as sio
 from itertools import cycle
 from omegaconf import DictConfig, OmegaConf
@@ -47,6 +48,7 @@ from sleap_nn.training.callbacks import (
     WandBPredImageLogger,
     CSVLoggerCallback,
 )
+from sleap_nn import GLOBAL_RANK, WORLD_SIZE
 
 
 @attrs.define
@@ -167,7 +169,7 @@ class ModelTrainer:
             )
             for label in labels:
                 train_split, val_split = label.make_training_splits(
-                    n_train=1 - val_fraction, n_val=val_fraction
+                    n_train=1 - val_fraction, n_val=val_fraction, seed=42
                 )
                 self.train_labels.append(train_split)
                 self.val_labels.append(val_split)
@@ -308,10 +310,13 @@ class ModelTrainer:
         # if save_ckpt_path is None, assign a new dir name
         ckpt_path = self.config.trainer_config.save_ckpt_path
         if ckpt_path is None:
-            ckpt_path = (
-                datetime.now().strftime("%y%m%d_%H%M%S")
-                + f".{self.model_type}.n={len(self.train_labels)+len(self.val_labels)}"
-            )
+            if WORLD_SIZE > 1:
+                ckpt_path = f".{self.model_type}.n={len(self.train_labels)+len(self.val_labels)}"
+            else:
+                ckpt_path = (
+                    datetime.now().strftime("%y%m%d_%H%M%S")
+                    + f".{self.model_type}.n={len(self.train_labels)+len(self.val_labels)}"
+                )
 
         self.config.trainer_config.save_ckpt_path = ckpt_path
 
@@ -355,18 +360,17 @@ class ModelTrainer:
                 logger.error(message)
                 raise OSError(message)
 
-        OmegaConf.save(
-            self._initial_config, (Path(ckpt_path) / "initial_config.yaml").as_posix()
+    def _setup_viz_datasets(self):
+        """Setup dataloaders."""
+        data_viz_config = self.config.copy()
+        data_viz_config.data_config.data_pipeline_fw = "torch_dataset"
+
+        return get_train_val_datasets(
+            train_labels=self.train_labels,
+            val_labels=self.val_labels,
+            config=data_viz_config,
+            rank=-1,
         )
-        for idx, (train, val) in enumerate(zip(self.train_labels, self.val_labels)):
-            train.save(
-                Path(ckpt_path) / f"labels_train_gt_{idx}.slp",
-                restore_original_videos=False,
-            )
-            val.save(
-                Path(ckpt_path) / f"labels_val_gt_{idx}.slp",
-                restore_original_videos=False,
-            )
 
     def _setup_datasets(self):
         """Setup dataloaders."""
@@ -401,10 +405,10 @@ class ModelTrainer:
             train_labels=self.train_labels,
             val_labels=self.val_labels,
             config=self.config,
-            rank=self.trainer.global_rank if self.trainer is not None else None,
+            rank=GLOBAL_RANK,
         )
 
-    def _setup_loggers_callbacks(self, train_dataset, val_dataset):
+    def _setup_loggers_callbacks(self, viz_train_dataset, viz_val_dataset):
         """Create loggers and callbacks."""
         logger.info("Setting up callbacks and loggers...")
         loggers = []
@@ -462,7 +466,8 @@ class ModelTrainer:
             if wandb_config.wandb_mode == "offline":
                 os.environ["WANDB_MODE"] = "offline"
             else:
-                wandb.login(key=self.config.trainer_config.wandb.api_key)
+                if GLOBAL_RANK in [0, -1]:
+                    wandb.login(key=self.config.trainer_config.wandb.api_key)
             wandb_logger = WandbLogger(
                 entity=wandb_config.entity,
                 project=wandb_config.project,
@@ -490,12 +495,13 @@ class ModelTrainer:
 
         # viz callbacks
         if self.config.trainer_config.visualize_preds_during_training:
-            train_viz_pipeline = cycle(train_dataset)
-            val_viz_pipeline = cycle(val_dataset)
+            train_viz_pipeline = cycle(viz_train_dataset)
+            val_viz_pipeline = cycle(viz_val_dataset)
 
             viz_dir = Path(self.config.trainer_config.save_ckpt_path) / "viz"
             if not Path(viz_dir).exists():
-                Path(viz_dir).mkdir(parents=True, exist_ok=True)
+                if GLOBAL_RANK in [0, -1]:
+                    Path(viz_dir).mkdir(parents=True, exist_ok=True)
 
             callbacks.append(
                 MatplotlibSaver(
@@ -517,8 +523,8 @@ class ModelTrainer:
             )
 
             if self.model_type == "bottomup":
-                train_viz_pipeline1 = cycle(copy.deepcopy(train_dataset))
-                val_viz_pipeline1 = cycle(copy.deepcopy(val_dataset))
+                train_viz_pipeline1 = cycle(copy.deepcopy(viz_train_dataset))
+                val_viz_pipeline1 = cycle(copy.deepcopy(viz_val_dataset))
                 callbacks.append(
                     MatplotlibSaver(
                         save_folder=viz_dir,
@@ -539,8 +545,8 @@ class ModelTrainer:
                 )
 
             if self.model_type == "multi_class_bottomup":
-                train_viz_pipeline1 = cycle(copy.deepcopy(train_dataset))
-                val_viz_pipeline1 = cycle(copy.deepcopy(val_dataset))
+                train_viz_pipeline1 = cycle(copy.deepcopy(viz_train_dataset))
+                val_viz_pipeline1 = cycle(copy.deepcopy(viz_val_dataset))
                 callbacks.append(
                     MatplotlibSaver(
                         save_folder=viz_dir,
@@ -612,14 +618,17 @@ class ModelTrainer:
         total_params = sum(p.numel() for p in self.lightning_model.parameters())
         self.config.model_config.total_params = total_params
 
-        # create the train and val datasets.
-        logger.info(f"Setting up train and val datasets...")
-        train_dataset, val_dataset = self._setup_datasets()
+        # create the train and val datasets for vizualization.
+        viz_train_dataset = None
+        viz_val_dataset = None
+        if self.config.trainer_config.visualize_preds_during_training:
+            logger.info(f"Setting up vizualization train and val datasets...")
+            viz_train_dataset, viz_val_dataset = self._setup_viz_datasets()
 
         # setup loggers and callbacks for Trainer.
         logger.info(f"Setting up Trainer...")
         loggers, callbacks = self._setup_loggers_callbacks(
-            train_dataset=train_dataset, val_dataset=val_dataset
+            viz_train_dataset=viz_train_dataset, viz_val_dataset=viz_val_dataset
         )
         # set up the strategy (for multi-gpu training)
         strategy = OmegaConf.select(
@@ -635,6 +644,26 @@ class ModelTrainer:
                 message = f"{cfg_profiler} is not a valid option. Please choose one of {list(self._profilers.keys())}"
                 logger.error(message)
                 raise ValueError(message)
+
+        # create lightning.Trainer instance.
+        self.trainer = L.Trainer(
+            callbacks=callbacks,
+            logger=loggers,
+            enable_checkpointing=self.config.trainer_config.save_ckpt,
+            devices=self.config.trainer_config.trainer_devices,
+            max_epochs=self.config.trainer_config.max_epochs,
+            accelerator=self.config.trainer_config.trainer_accelerator,
+            enable_progress_bar=self.config.trainer_config.enable_progress_bar,
+            strategy=strategy,
+            profiler=profiler,
+            log_every_n_steps=1,
+        )  # TODO check any other methods to use rank in dataset creations!
+
+        self.trainer.strategy.barrier()
+
+        # setup datasets
+        train_dataset, val_dataset = self._setup_datasets()
+
         # set-up steps per epoch
         train_steps_per_epoch = self.config.trainer_config.train_steps_per_epoch
         if train_steps_per_epoch is None:
@@ -651,20 +680,7 @@ class ModelTrainer:
             batch_size=self.config.trainer_config.val_data_loader.batch_size,
         )
 
-        # create lightning.Trainer instance.
-        self.trainer = L.Trainer(
-            callbacks=callbacks,
-            logger=loggers,
-            enable_checkpointing=self.config.trainer_config.save_ckpt,
-            devices=self.config.trainer_config.trainer_devices,
-            max_epochs=self.config.trainer_config.max_epochs,
-            accelerator=self.config.trainer_config.trainer_accelerator,
-            enable_progress_bar=self.config.trainer_config.enable_progress_bar,
-            limit_train_batches=self.config.trainer_config.train_steps_per_epoch,
-            strategy=strategy,
-            profiler=profiler,
-            log_every_n_steps=1,
-        )  # TODO check any other methods to use rank in dataset creations!
+        self.trainer.limit_train_batches = train_steps_per_epoch
 
         # setup dataloaders
         # need to set up dataloaders after Trainer is initialized (for ddp). DistributedSampler depends on the rank
@@ -672,14 +688,26 @@ class ModelTrainer:
             train_dataset=train_dataset,
             val_dataset=val_dataset,
             config=self.config,
-            rank=self.trainer.global_rank if self.trainer is not None else None,
+            rank=self.trainer.global_rank,
             train_steps_per_epoch=self.config.trainer_config.train_steps_per_epoch,
             val_steps_per_epoch=val_steps_per_epoch,
         )
 
-        if (
-            self.trainer.global_rank == 0
-        ):  # save config if there are no distributed process
+        if self.trainer.global_rank == 0:  # save config only in rank 0 process
+            ckpt_path = self.config.trainer_config.save_ckpt_path
+            OmegaConf.save(
+                self._initial_config,
+                (Path(ckpt_path) / "initial_config.yaml").as_posix(),
+            )
+            for idx, (train, val) in enumerate(zip(self.train_labels, self.val_labels)):
+                train.save(
+                    Path(ckpt_path) / f"labels_train_gt_{idx}.slp",
+                    restore_original_videos=False,
+                )
+                val.save(
+                    Path(ckpt_path) / f"labels_val_gt_{idx}.slp",
+                    restore_original_videos=False,
+                )
 
             if self.config.trainer_config.use_wandb:
                 if wandb.run is None:
@@ -705,6 +733,8 @@ class ModelTrainer:
                 ).as_posix(),
             )
 
+        self.trainer.strategy.barrier()
+
         try:
             logger.info(
                 f"Finished trainer set up. [{time.time() - start_setup_time:.1f}s]"
@@ -725,7 +755,7 @@ class ModelTrainer:
             logger.info(
                 f"Finished training loop. [{(time.time() - start_train_time) / 60:.1f} min]"
             )
-            if self.config.trainer_config.use_wandb:
+            if self.trainer.global_rank == 0 and self.config.trainer_config.use_wandb:
                 wandb.finish()
 
             # delete image disk caching
@@ -734,14 +764,16 @@ class ModelTrainer:
                 == "torch_dataset_cache_img_disk"
                 and self.config.data_config.delete_cache_imgs_after_training
             ):
-                self._delete_cache_imgs()
+                if self.trainer.global_rank == 0:
+                    self._delete_cache_imgs()
 
             # delete viz folder if requested
             if (
                 self.config.trainer_config.visualize_preds_during_training
                 and not self.config.trainer_config.keep_viz
             ):
-                viz_dir = Path(self.config.trainer_config.save_ckpt_path) / "viz"
-                if viz_dir.exists():
-                    logger.info(f"Deleting viz folder at {viz_dir}...")
-                    shutil.rmtree(viz_dir, ignore_errors=True)
+                if self.trainer.global_rank == 0:
+                    viz_dir = Path(self.config.trainer_config.save_ckpt_path) / "viz"
+                    if viz_dir.exists():
+                        logger.info(f"Deleting viz folder at {viz_dir}...")
+                        shutil.rmtree(viz_dir, ignore_errors=True)
